In photo video support

+++ src/main/java/cn/inphoto/task/WebInfoTask.java
@@ -16,7 +16,6 @@ import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.Resource;
-import java.io.IOException;
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Date;
@@ -128,23 +127,30 @@ public class WebInfoTask {
                 List<OSSObjectSummary> sums = objectListing.getObjectSummaries();
                 StringBuilder a = new StringBuilder();
 
-                for (OSSObjectSummary s : sums) {
-                    boolean flag = false;
-                    for (MediaData m : mediaDataList
-                            ) {
-                        if (s.getKey().equals(m.getMediaKey())) {
-                            flag = true;
-                            break;
+                if (sums != null && !sums.isEmpty()) {
+                    for (OSSObjectSummary s : sums) {
+                        boolean flag = false;
+
+                        for (MediaData m : mediaDataList
+                                ) {
+                            try {
+                                if (s.getKey().equals(m.getMediaKey())) {
+                                    flag = true;
+                                    break;
+                                }
+                            } catch (Exception e) {
+                                logger.log(UserLogLevel.TASK, "清理无效的设置文件,media_id=" + m.getMediaId() + "的媒体数据key为空");
+                            }
                         }
-                    }
-                    if (!flag) {
-                        isDelete = true;
-                        MediaData md = mediaDataDao.findByMediaKey(s.getKey());
-                        md.setMediaState(MediaData.MediaState.Delete);
-                        md.setDeleteTime(new Timestamp(new Date().getTime()));
-                        md.setOverTime(new Timestamp(new Date().getTime()));
-                        client.deleteObject(bucketName, s.getKey());
-                        a.append(s.getKey()).append("、");
+                        if (!flag) {
+                            isDelete = true;
+                            MediaData md = mediaDataDao.findByMediaKey(s.getKey());
+                            md.setMediaState(MediaData.MediaState.Delete);
+                            md.setDeleteTime(new Timestamp(new Date().getTime()));
+                            md.setOverTime(new Timestamp(new Date().getTime()));
+                            client.deleteObject(bucketName, s.getKey());
+                            a.append(s.getKey()).append("、");
+                        }
                     }
                 }
